[k] Batch fixes 20260913 (#6340)
* feat(kunlunxin): batch3 kunlunxin fixes on upstream master baseline

Batch 3 verification closure (303 ops): 207 files on top of
upstream/master fcdd6ee97.
- 163 PASS / 129 BLOCKED / 11 SKIP (dtype-equal-weight Gems Speedup >= 0.8 gate)
- Vendor-dir changes under _kunlunxin/ (benchmark/ and tests/ changes
  split onto batch3_tests_benchmark_20260915)
- No CPU/ATen/native fallback used as fix or PASS basis


* feat(kunlunxin): remove flagtune/hygon/mthreads/thead expand files from batch3

Remove 6 files added in batch3 commit that are not part of the kunlunxin
backend fixes: flagtune cli/package.py, cli/rebench.py,
contracts/configs/mul_flagtune_configs.yaml, and the expand yaml files
under _hygon/_mthreads/_thead.


* feat(kunlunxin): batch3 fixes - missing aten entries, tle.raw fast paths, trim vendor notes

- add missing out-of-place / in-place entries (divide, true_divide_tensor_,
  subtract, scatter_add, special_log1p) so direct dispatches reach the
  vendor kernels
- add tle.raw tensor-vs-scalar fast paths for greater/not_equal
  (gt_raw.xpu / ne_raw.xpu payloads, P800 cluster)
- fix tle.raw dialect name xpu3 -> xpu in lt/masked_fill/nan_to_num
- add is_scatter_slice to CodeGenConfig
- drop large header notes from vendor files (mhc_pre, add_rms_norm,
  bf16_paged_mqa_logits, bin_topk, matmuladd, mean)


* feat(kunlunxin): batch3 fixes - drop remaining ATen fallbacks in vendor ops

- smooth_l1_loss: replace torch.sum on the broadcast path with a two-stage
  fp32 triton reduce (mean folded into stage-1), reusing the final-sum kernel
- linalg_svdvals: host-side numpy descending sort instead of device-side
  vendor torch.sort (norms already live on the host; avoids radix launches)
- linalg_ldl_factor_ex: zero the strict upper triangle in-kernel on the LD
  store (tl.where value mask) instead of host torch.tril
- hadamard_transform: F.pad -> torch.zeros + copy_ (no ATen constant_pad_nd)
- segment_reduce: remove torch squeeze/sum/mean/amax/amin/prod fallback block
- nllloss: nll_loss2d_forward -> torch.autograd.Function routing fwd/bwd
  straight to the vendor kernels (autograd edges preserved)
- copy: drop dead ATen redispatch in the empty-tensor path; mode: drop dead
  torch.nn.functional import; conv_transpose1d / mkldnn_rnn_layer /
  fused_recurrent_gated_delta_rule_fwd: remove vendor overrides (native ATen)
- _copy_from: keep the raw ATen copy on the comparison kernels (tle dsa
  fp32->bool path measured 100x+ slower); trim leftovers


* chore(kunlunxin): trim vendor comments and license headers across backend

Clean up per-file docstrings/comments in the kunlunxin vendor backend
(ops/, fused/, utils/): keep one-line functional docstrings, drop
multi-line vendor notes, timing history, and license header blocks.


* style(kunlunxin): apply black/isort formatting across vendor backend

Reformat the kunlunxin vendor backend (ops/, fused/, utils/): wrap
long signatures and expressions, sort/reorder imports, and drop
stray blank lines. No functional changes.


* chore(kunlunxin): remove unused imports, dead code, and duplicates

- sub.py: drop duplicate subtract definition (already defined once)
- ops/__init__.py: drop duplicate bottom expand_copy import, unused torch
- linalg_solve_triangular / linalg_matrix_norm / median / nonzero_static /
  bin_topk / max_pool3d_with_indices: remove unused variables/loads
- layernorm / logical_not / scatter_add_ / sort: remove unused imports
- fused+ops __init__: annotate side-effect imports with noqa F401
- hc_head_fused_kernel: use importlib.import_module for vllm_xpu._C
- core_shapes.yaml: drop duplicated smooth_l1_loss_backward key

No functional changes.


* fix(kunlunxin): replace torch.matmul with vendor bmm in linalg_svd

The one-sided Jacobi U^H A product in _osj_svd_impl now goes through
the kunlunxin bmm kernel instead of the ATen matmul fallback.


* fix(kunlunxin): restore upstream _FULL_CONFIG registrations, drop _linalg_eigvals shim

- flag_gems/__init__.py: restore the 24 upstream _FULL_CONFIG entries that the
  batch tree commit (28483c69) reverted from an older base; the deletions
  survived the later master merges.  Restores cov/corrcoef/nuclear_norm/
  linalg_matrix_rank (x8)/hash_tensor/sym_size/slice.Tensor/dsplit/gru/
  embedding_renorm_/cumulative_trapezoid.dx/cosine_embedding_loss/
  special_hermite_polynomial_h.n_scalar/_fake_quantize_* and the upstream
  adaptive_avg_pool3d_backward dispatch fix (749a6d2a).
- _kunlunxin: drop the _linalg_eigvals XPU redispatch shim (remove the module,
  its import and __all__ entry) and list it in CUSTOMIZED_UNUSED_OPS.
- linalg_matrix_norm: clamp+sqrt in-kernel in _sturm_eig_kernel so the caller
  no longer needs torch.clamp/torch.sqrt on the eigenvalues; drop the stale
  header docstring.
- linalg_svdvals: drop the stale header docstring, add the license header to
  pointwise_dynamic.


* fix(kunlunxin): device-side pad/sort kernels, drop _copy_from fallbacks

- linalg_matrix_norm: add `_pad_rows_kernel` so padding a non-BN-multiple
  column count runs on device, and call `tle_copy` directly instead of
  falling back to `aten::_copy_from` on the strided copies.
- linalg_svdvals: replace the host numpy descending sort of the singular
  values with a device-side bitonic sort stage kernel.
- __init__: register the remaining linalg_eig / linalg_eigvals variants in
  CUSTOMIZED_UNUSED_OPS.
- reglu: drop stale probe-tuning docstrings (no functional change).

* fix(kunlunxin): normalize GEMS_KUNLUNXIN debug logs, fuse bin_topk select

- Every vendor logger.debug call now uses the `GEMS_KUNLUNXIN <OP>` prefix;
  13 sites across 11 files still logged `GEMS <OP>` or `GEMS KUNLUNXIN ...`.
- bin_topk: fold `cands = thrs | (1 << bit)` and the trailing `torch.where`
  of the radix-select loop into `_bst_count_kernel`, which now updates
  `thrs` in place.
- linalg_matrix_norm: lay `_pad_rows_kernel` out as a single flat 1D grid
  over a buffer sized in whole row blocks (a non-affine store index costs
  5x: 365us -> 32us for an 8x128 buffer).
- linalg_svdvals: rename the bitonic loop variable `l` -> `lev` (flake8
  E741) and expand the stage-kernel signature.

* perf(kunlunxin): cheaper padding and device sort

- linalg_matrix_norm: fill the pad buffer with native `torch.full` (1187GB/s
  vs ~29GB/s for a kernel writing the whole buffer) and only touch the live
  rows -- `_pad_col_kernel` scatters the C == 1 case (the one destination
  `tle_copy` cannot express) while C > 1 stays on `tle_copy` and now fails
  loudly instead of silently reducing an all-identity buffer.  Both kernels
  keep the clamp on the load, since a non-affine store index costs 5x here.
- linalg_svdvals: launch the bitonic stage kernel one program per (stage,
  row) with the direction patterns packed into a single host-built table --
  shifted-affine kernels are silently corrupted when a launch runs more than
  one program or one program runs more than one stage.

---------

+++ src/flag_gems/runtime/backend/_kunlunxin/ops/dequantize.py
@@ -1,8 +1,3 @@
-# Copyright 2026 FlagOS Contributors
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-
 import logging
 
 import torch
@@ -13,26 +8,12 @@ from flag_gems.runtime import torch_device_fn
 
 logger = logging.getLogger(__name__)
 
-# Element-size preserving integer alias for every supported quantized dtype.
-# quint8 is aliased through int8 on purpose: an unsigned->float conversion
-# (arith.uitofp) is mis-lowered by TritonXPU for BLOCK >= 256, so the raw bytes
-# are read as signed and folded back into [0, 255] inside the kernel.
 _QINT_ALIAS = {
     torch.qint8: (torch.int8, False),
     torch.quint8: (torch.int8, True),
     torch.qint32: (torch.int32, False),
 }
 
-# Vector stores on this backend always touch a full 64-element granule and a
-# masked store is not honoured, so the destination buffer is padded up to a
-# whole tile (BLOCK is a multiple of 64) and the kernel stores without a mask.
-# The float32 view handed back to the caller only covers the real elements.
-#
-# One tile per program (a grid-stride loop measured 1.44 GB/s against 209 GB/s
-# for the flat form) and BLOCK=8192, the top of the measured bandwidth curve
-# (512 -> 41.8, 1024 -> 73.0, 2048 -> 116.7, 4096 -> 166.0, 8192 -> 209.6,
-# 16384/32768 -> 208.3 GB/s at 16.7M elements). num_warps has no measurable
-# effect here, so the default is kept.
 _BLOCK = 8192
 
 
@@ -47,11 +28,8 @@ def _dequantize_kernel(
     BLOCK: tl.constexpr,
 ):
     offsets = tl.program_id(0) * BLOCK + tl.arange(0, BLOCK)
-    # No `other=` here: on TritonXPU the `other` operand silently clobbers
-    # valid lanes. Lanes past n_elements read undefined bytes and land in
-    # the padded tail of `out_ptr`, which the caller never exposes.
     raw = tl.load(x_ptr + offsets, mask=offsets < n_elements)
-    values = raw.to(tl.float32)
+    values = raw.to(tl.int32).to(tl.float32)
     if UNSIGNED:
         values = tl.where(values < 0.0, values + 256.0, values)
     tl.store(out_ptr + offsets, (values - zero_point) * scale)
